Updated float precision

+++ pkg/drivers/MediaInfo/MediaInfoDriver.go
@@ -107,16 +107,16 @@ func mapMediaInfoJsonToVideoMetaData(mediaInfoOutput MediaInfoOutput, videoMetad
 			videoMetadata.Height = utils.ParseInt(track.Height)
 			videoMetadata.Width = utils.ParseInt(track.Width)
 			videoMetadata.AspectRatio = utils.AspectRatioFloatToString(videoMetadata.Width, videoMetadata.Height)
-			videoMetadata.VideoDuration = utils.ParseFloat64(track.Duration)
+			videoMetadata.VideoDuration = utils.ParseFloat32(track.Duration)
 			videoMetadata.VideoCodec = track.Format
-			videoMetadata.VideoBitRate = utils.ParseFloat64(track.BitRate)
-			videoMetadata.VideoFrameRate = utils.ParseFloat64(track.FrameRate)
+			videoMetadata.VideoBitRate = utils.ParseFloat32(track.BitRate)
+			videoMetadata.VideoFrameRate = utils.ParseFloat32(track.FrameRate)
 		case TYPE_AUDIO:
 			videoMetadata.AudioCodec = track.Format
-			videoMetadata.AudioBitRate = utils.ParseFloat64(track.BitRate)
-			videoMetadata.AudioFrameRate = utils.ParseFloat64(track.FrameRate)
+			videoMetadata.AudioBitRate = utils.ParseFloat32(track.BitRate)
+			videoMetadata.AudioFrameRate = utils.ParseFloat32(track.FrameRate)
 			videoMetadata.AudioChannels = utils.ParseInt(track.Channels)
-			videoMetadata.AudioDuration = utils.ParseFloat64(track.Duration)
+			videoMetadata.AudioDuration = utils.ParseFloat32(track.Duration)
 		}
 	}
 }
